requirements.txt
streamlit
praat-parselmouth
numpy
matplotlib